Inform about breaking change
The `Handle ActiveRecord transaction correctly, not storing or removing files on rollback` should be added as a breaking change. 

We discoverd a bug in our application due to this.

Since it changed a callback from `after_save` to `after_commit`, you must re-order your own `after_commit` callbacks (placing them _before_  the gem's `mount_uploader` macro) if you want them to be run _after_  the store is performed.

Otherwise (if your own `after_commit` is declared _after_  the `mount_uploader` call), Rails will actually run it before the gem's `after_commit` (it's weird, but it's a known bug in Rails 5 and still present in Rails 6).